Clearing and Relocating Files
