--- src/strategy.js
@@ -1,72 +1,0 @@
-//BASE DO PROJETO 
-
-class NotImplementedException extends Error {
-    constructor() {
-        super("Not Implemented Exception");
-    }
-};
-
-//Interface
-class ICrud {//Crud quando não implementado
-    create(item){
-        throw new NotImplementedException();
-    };
-
-    read(query) {
-        throw new NotImplementedException();
-    };
-
-    update(id, item) {
-        throw new NotImplementedException();
-    };
-
-    delete(id) {
-        throw new NotImplementedException();
-    };
-};
-
-class MongoDB extends ICrud{
-    constructor() {
-        super();
-    };
-
-    create(item) {
-        console.log("O item foi salvo em MongoDB");
-    }
-};
-
-class Postgres extends ICrud {
-    constructor() {
-        super()
-    };
-
-    create(item) {
-        console.log("O item foi salvo em Postgres");
-    }
-};
-
-class ContextStrategy extends ICrud {//Classe abstrata
-    constructor(strategy){
-        super();
-        this._database = strategy;
-    }
-
-    create(item) {
-        return this._database.create(item);
-    };
-    read(item) {
-        return this._database.read(item);
-    };
-    update(id, item) {
-        return this._database.update(id, item);
-    };
-    delete(id) {
-        return this._database.delete(id);
-    }
-};
-
-const contextMongo = new ContextStrategy(new MongoDB());
-contextMongo.create();
-
-const contextProstgres = new ContextStrategy(new Postgres());
-contextProstgres.create()
