+++ src/tests/apiHeros.test.js
@@ -1,5 +1,5 @@
 const assert = require('assert');
-const api = require('./../api');
+const api = require('./../../api');
 let app = {};
 const TOKEN = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6Impvc2VwIiwiaWQiOjEsImlhdCI6MTU0ODk0OTE0OX0.uYuurCcxJRGJwuY0cW85G6UlFnXr5B85-LWMmaJ3QGc'
 const headers = {
